Fixed for account of new prototype of buildPeerMessage method
The method Router2RouterPeer::buildPeerMsg that received a reference to
a PeerMessage has been changed to allocate itself the memory for
PeerMessage (so that now PeerMessage is always dynamically allocated),
and to return the pointer to the allocated message.
In this way, now the allocation and destroy of the PeerMessages is only
done in the Router2RouterPeer class.

git-svn-id: https://euclid.esac.esa.int/svn/ESA/SOC/SOC-3-DEV/SOC-3-07-QLook/QPF/trunk@11924 463500fe-d6be-4184-9b20-47717e75ecac